Fix outside package target activities not opening
This fix includes any launch URLs set to open the browser or another app

The issue was due to an Activity backstack not being built correctly.
Both `NotificationOpenedReceiver` and the target `Activity` were
launched with `FLAG_ACTIVITY_NEW_TASK` however this means when
`NotificationOpenedReceiver` is finished it does not bring up the target
since it is on a different task. The Android flag `allowTaskReparenting`
allows this to happen, however this can effect the backstack of the main
task which we don't want to do.

Setting `android:taskAffinity` does `FLAG_ACTIVITY_NEW_TASK` and
`allowTaskReparenting` for us. By setting the value to "" (empty string)
it prevents any side effects on existing tasks as it means it has no
affinity to associate it with. This allowed us to clean up some runtime
logic to add `FLAG_ACTIVITY_NEW_TASK` and others we no longer need now.