Fixing rounded corners for IAMs
cardViews have their own custom background where corner radius is applied. By setting the backgroundColor you remove the custom background and replace it with a new one that has the new color. To avoid this we should be using the setCardBackgroundColor instead. This changes the color on the custom background instead of replacing it.