rm onesignal-gradle-plugin so substitution works
The onesignal-gradle-plugin currently breaks the Gradle
`dependencySubstitution` feature we introduced in the last commit so we
removed it from the example project.

Adding`googleServices { disableVersionCheck = true }`
was required since the `onesignal-gradle-plugin` auto disables this
part of the `com.google.gms.google-services` for us. The current
OneSignal-Android-SDK doesn't work with it since it uses version ranges
which is a re-existing issue.