Use GROUP_ALERT_CHILDREN to fix heads up summaries
Using GROUP_ALERT_CHILDREN with setGroupAlertBehavior means the just
receive notification will be used as the heads-up, if the notification
channel is configured for highest display priority. Before with
GROUP_ALERT_SUMMARY, you would see a summary of all notifications in the
group which was much harder to see the most recent.

This also fixes a bug where the previous notification would be shown as
the heads-up notification on some devices. Such as Samsung's OneUI 4
with the Brief style of notifications, as well as some other
manufacturers.
Lastly the bug would sometimes happen on Android 7 and 8 AOSP since the
summary and child notification are posted at the same time and these
versions of Android would pick one at random. In this case it seemed to
only be an issue when setting android_group and only on the 2nd
notification display.